notification-service/requirements.txt
redis==5.0.1
aiohttp==3.9.1
python-telegram-bot==20.7
python-dotenv==1.0.0